Convert ProcessorElement to have a backend
