+++ mfp/gui/clutter/__init__.py
@@ -6,3 +6,4 @@ from .text_widget import ClutterTextWidgetImpl  # noqa
 from .base_element import ClutterBaseElementBackend  # noqa
 from .colordb import ClutterColorDBBackend  # noqa
 from .message_element import ClutterMessageElementImpl  # noqa
+from .processor_element import ClutterProcessorElementImpl  # noqa
