+++ mfp/gui/processor_element.py
@@ -4,16 +4,21 @@ processor_element.py
 A patch element corresponding to a signal or control processor
 '''
 
-from gi.repository import Clutter
-import cairo
+from abc import ABC, abstractmethod
 from .text_widget import TextWidget
-from .colordb import ColorDB
 from .modes.label_edit import LabelEditMode
 from ..gui_main import MFPGUI
+from .backend_interfaces import BackendInterface
+from .base_element import BaseElement
 
-from .clutter.base_element import ClutterBaseElementBackend
 
-class ProcessorElement (ClutterBaseElementBackend):
+class ProcessorElementImpl(ABC, BackendInterface):
+    @abstractmethod
+    def redraw(self):
+        pass
+
+
+class ProcessorElement (BaseElement):
     display_type = "processor"
     proc_type = None
 
@@ -21,7 +26,10 @@ class ProcessorElement (ClutterBaseElementBackend):
     label_off_x = 3
     label_off_y = 0
 
-    def __init__(self, window, x, y, params={}):
+    def __init__(self, window, x, y, params=None):
+        if params is None:
+            params = {}
+
         super().__init__(window, x, y)
 
         self.param_list.extend([
@@ -34,9 +42,17 @@ class ProcessorElement (ClutterBaseElementBackend):
         self.show_label = params.get("show_label", True)
 
         # display elements
-        self.texture = None
-        self.label = None
+        self.label = TextWidget.get_factory()(self)
+        self.label.set_position(self.label_off_x, self.label_off_y)
+        self.label.set_color(self.get_color('text-color'))
+        self.label.set_font_name(self.get_fontspec())
+        self.label.signal_listen('text-changed', self.label_changed_cb)
+        self.label.set_reactive(False)
         self.label_text = None
+
+        if not self.show_label:
+            self.label.hide()
+
         self.export_x = None
         self.export_y = None
         self.export_w = None
@@ -44,33 +60,11 @@ class ProcessorElement (ClutterBaseElementBackend):
         self.export_created = False
 
         # create display
-        self.create_display()
         self.set_size(35, 25)
         self.move(x, y)
 
         self.obj_state = self.OBJ_HALFCREATED
-
         self.update()
-
-    def create_display(self):
-        # box
-        self.texture = Clutter.Canvas.new()
-        self.group.set_content(self.texture)
-        self.texture.connect("draw", self.draw_cb)
-        self.texture.set_size(35, 25)
-
-        # label
-        self.label = TextWidget.get_factory()(self)
-        self.label.set_position(self.label_off_x, self.label_off_y)
-        self.label.set_color(self.get_color('text-color'))
-        self.label.set_font_name(self.get_fontspec())
-        self.label.signal_listen('text-changed', self.label_changed_cb)
-        self.label.set_reactive(False)
-
-        if not self.show_label:
-            self.label.hide()
-
-        self.group.set_reactive(True)
 
     def update(self):
         if self.show_label or self.obj_state == self.OBJ_HALFCREATED:
@@ -88,45 +82,6 @@ class ProcessorElement (ClutterBaseElementBackend):
         new_w = max(35, port_width, label_width, box_width)
 
         self.set_size(new_w, self.texture.get_property('height'))
-
-    def draw_cb(self, texture, ct, width, height):
-        lw = 2.0
-
-        w = width - lw
-        h = height - lw
-
-        # clear the drawing area
-        ct.save()
-        ct.set_operator(cairo.OPERATOR_CLEAR)
-        ct.paint()
-        ct.restore()
-
-        ct.set_line_width(lw)
-        ct.set_antialias(cairo.ANTIALIAS_NONE)
-        ct.translate(lw/2.0, lw/2.0)
-        ct.move_to(0, 0)
-        ct.line_to(0, h)
-        ct.line_to(w, h)
-        ct.line_to(w, 0)
-        ct.line_to(0, 0)
-        ct.close_path()
-
-        # fill to paint the background
-        color = ColorDB().normalize(self.get_color('fill-color'))
-        ct.set_source_rgba(color.red, color.green, color.blue, color.alpha)
-        ct.fill_preserve()
-
-        # stroke to draw the outline
-        color = ColorDB().normalize(self.get_color('stroke-color'))
-        ct.set_source_rgba(color.red, color.green, color.blue, color.alpha)
-
-        if self.obj_state == self.OBJ_COMPLETE:
-            ct.set_dash([])
-        else:
-            ct.set_dash([8, 4])
-        ct.set_line_width(lw)
-        ct.stroke()
-        return True
 
     def get_label(self):
         return self.label
@@ -159,28 +114,6 @@ class ProcessorElement (ClutterBaseElementBackend):
             self.label.hide()
 
         self.update()
-
-    def label_changed_cb(self, *args):
-        newtext = self.label.get_text()
-        if newtext != self.label_text:
-            self.label_text = newtext
-            self.update()
-
-    def set_size(self, w, h):
-        super().set_size(w, h)
-
-        self.texture.set_size(w, h)
-        self.texture.invalidate()
-
-    def select(self):
-        super().select()
-        self.label.set_color(self.get_color('text-color'))
-        self.texture.invalidate()
-
-    def unselect(self):
-        super().unselect()
-        self.label.set_color(self.get_color('text-color'))
-        self.texture.invalidate()
 
     async def make_edit_mode(self):
         return LabelEditMode(self.app_window, self, self.label)
@@ -227,4 +160,3 @@ class ProcessorElement (ClutterBaseElementBackend):
 
         if need_update:
             self.update()
-
